+++ SensumUdred/src/Data/IndividualData.java
@@ -17,19 +17,19 @@ public class IndividualData implements IIndividual{
 
     private String name;
     private String address;
-    private int CPR;
+    private long CPR;
     
     IndividualData(){
         
     }
 
-    IndividualData(String name, String address, int CPR) {
+    IndividualData(String name, String address, long CPR) {
         this.name = name;
         this.address = address;
         this.CPR = CPR;
     }
 
-    void setAttributes(String name, String address, int CPR){
+    void setAttributes(String name, String address, long CPR){
         this.name = name;
         this.address = address;
         this.CPR = CPR;
@@ -46,7 +46,7 @@ public class IndividualData implements IIndividual{
     }
 
     @Override
-    public int getCPR() {
+    public long getCPR() {
         return CPR;
     }
